Methods list

PERK.PERK

PERK.EuclideanKernel

PERK.EuclideanKernel

PERK.ExactKernel

PERK.ExactTrainingData

PERK.GaussianKernel

PERK.GaussianKernel

PERK.GaussianKernel

PERK.GaussianRFF

PERK.GaussianRFF

PERK.GaussianRFF

PERK.Kernel

PERK.RFFKernel

PERK.RFFTrainingData

PERK.TrainingData

PERK.addnoise!

PERK.combine

PERK.complex2real

PERK.div0

PERK.gaussiankernel

PERK.generatenoisydata

PERK.holdout

PERK.krr

PERK.krr_train

PERK.perk

PERK.perk

PERK.rffmap

PERK.train

Methods usage

PERK.PERK — Module

PERK

Module implementing parameter estimation via regression with kernels (PERK).

Exports

EuclideanKernel: Euclidean inner product kernel (for ridge regression instead of kernel ridge regression)

GaussianKernel: Gaussian kernel used in kernel ridge regression

GaussianRFF: Approximation of Gaussian kernel using random Fourier features

generatenoisydata: Function for generating noisy data

perk: Function for running PERK

PERK.EuclideanKernel — Type

EuclideanKernel() <: ExactKernel

Create a kernel function where the kernel is just the Euclidean inner product (i.e., ridge regression).

PERK.EuclideanKernel — Method

(k::EuclideanKernel)(p, q)

Evaluate the Euclidean inner product between p and q.

Arguments

p::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: First kernel input [Q,M] or [M] (if Q = 1) or scalar (if Q = M = 1)

q::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Second kernel input [Q,N] or [N] (if Q = 1) or scalar (if Q = N = 1)

Note

Q is the number of features

M is the number of feature vectors in the first input

N is the number of feature vectors in the second input

Return

K::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Kernel output [M,N] or [M] (if N = 1) or [N] (if M = 1) or scalar (if M = N = 1)

PERK.ExactKernel — Type

ExactKernel <: Kernel

Abstract type for representing kernels that are evaluated exactly. ExactKernels must be callable with two inputs.

PERK.ExactTrainingData — Type

ExactTrainingData(y, x, xm, K, Km, xKinv) <: TrainingData

Create an object that contains the training data when using the full Gram matrix K.

Properties

y::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Features for training data [Q,T] or [T] (if Q = 1)

x::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Latent parameters for training data [L,T] or [T] (if L = 1)

xm::Union{<:Real,<:AbstractVector{<:Real}}: Mean of latent parameters [L] or scalar (if L = 1)

K::AbstractMatrix{<:Real}: De-meaned (both rows and columns) Gram matrix of the kernel evaluated on the training data features [T,T]

Km::AbstractVector{<:Real}: Row means of K (before de-meaning) [T]

xKinv::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: x times the regularized inverse of K [L,T] or [T] (if L = 1)

Q::Integer: Number of training features

L::Integer: Number of latent parameters

T::Integer: Number of training points

PERK.GaussianKernel — Type

GaussianKernel(Λ) <: ExactKernel

Create a Gaussian kernel function.

Properties

Λ::Union{<:Real,AbstractVector{<:Real}}: Length scales [Q] or scalar (if Q = 1)

Note

Q is the number of features

PERK.GaussianKernel — Method

(k::GaussianKernel)(p, q)

Evaluate the Gaussian kernel.

Arguments

p::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: First kernel input [Q,M] or [M] (if Q = 1) or scalar (if Q = M = 1)

q::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Second kernel input [Q,N] or [N] (if Q = 1) or scalar (if Q = N = 1)

Note

Q is the number of features

M is the number of feature vectors in the first input

N is the number of feature vectors in the second input

Return

K::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Kernel output [M,N] or [M] (if N = 1) or [N] (if M = 1) or scalar (if M = N = 1)

PERK.GaussianKernel — Method

GuassianKernel(Λy, [Λν])

Create a Guassian kernel function.

Arguments

Λy::Union{<:Number,<:AbstractVector{<:Number}}: Length scales for features [Q] or scalar (if Q = 1)

Λν::Union{<:Real,<:AbstractVector{<:Real}}: Length scales for known parameters [K] or scalar (if K = 1)

PERK.GaussianRFF — Type

GaussianRFF(H, Λ) <: RFFKernel

Create an approximate (via random Fourier features) Gaussian kernel function.

Properties

H::Integer: Approximation order

Λ::Union{<:Real,AbstractVector{<:Real}}: Length scales [Q] or scalar (if Q = 1)

Note

Q is the number of features

PERK.GaussianRFF — Method

GuassianRFF(H, Λy, [Λν])

Create an approximate (via random Fourier features) Gaussian kernel function.

Arguments

H::Integer: Approximation order

Λy::Union{<:Number,<:AbstractVector{<:Number}}: Length scales for features [Q] or scalar (if Q = 1)

Λν::Union{<:Real,<:AbstractVector{<:Real}}: Length scales for known parameters [K] or scalar (if K = 1)

PERK.GaussianRFF — Method

(k::GaussianRFF)([rng], q)
(k::GaussianRFF)(q, f, phase)

Evaluate the approximate Gaussian kernel.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use to generate f and phase

q::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Kernel input [Q,N] or [N] (if Q = 1) or scalar (if Q = N = 1)

f::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}} = randn(k.H, Q): Unscaled random frequency values [H,Q] or [H] (if Q = 1)

phase::AbstractVector{<:Real} = rand(k.H): Random phase values [H]

Note

Q is the number of features

N is the number of feature vectors in the input

H is the approximation order for the random Fourier features

Return

z::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Higher-dimensional features [H,N] or [H] (if N = 1)

freq::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Random frequency values [H,Q] or [H] (if Q = 1)

phase::AbstractVector{<:Real}: Random phase values [H]

PERK.Kernel — Type

Kernel

Abstract type for representing kernel functions.

PERK.RFFKernel — Type

RFFKernel <: Kernel

Abstract type for representing kernels that are approximated via random Fourier features. RFFKernels must be callable with one or three inputs.

PERK.RFFTrainingData — Type

RFFTrainingData(freq, phase, zm, xm, Czz, Cxz, CxzCzzinv) <: TrainingData

Create an object that contains the training data when using an approximation of the Gram matrix K using random Fourier features.

Properties

freq::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Random frequency values for random Fourier features [H,Q] or [H] (if Q = 1)

phase::AbstractVector{<:Real}: Random phase values for random Fourier features [H]

zm::AbstractVector{<:Real}: Mean of feature maps [H]

xm::Union{<:Real,<:AbstractVector{<:Real}}: Mean of latent parameters [L] or scalar (if L = 1)

Czz::AbstractMatrix{<:Real}: Auto-covariance matrix of feature maps [H,H]

Cxz::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Cross-covariance matrix between latent parameters and feature maps [L,H] or [H] (if L = 1)

CxzCzzinv::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Cxz times the regularized inverse of Czz [L,H] or [H] (if L = 1)

Q::Integer: Number of training features

L::Integer: Number of latent parameters

H::Integer: Kernel approximation order

PERK.TrainingData — Type

TrainingData

Abstract type for representing training data.

PERK.addnoise! — Method

addnoise!([rng], y, noiseDist)

Add noise to y. If elements of y are complex-valued, then add independent noise to both the real and imaginary parts of y.

PERK.combine — Method

combine(y, ν)

Combine the output of the signal models with the known parameters.

PERK.complex2real — Method

complex2real(y)

Split complex data into real and imaginary parts.

PERK.div0 — Method

div0(a, b)

Compute a / b, but return 0 if b is 0.

PERK.gaussiankernel — Method

gaussiankernel(p, q, sqrtΣ)

Compute the Gaussian kernel with covariance matrix Σ evaluated at p and q. Note that function input is sqrtΣ, i.e., the square root of Σ.

PERK.generatenoisydata — Method

generatenoisydata([rng], N, xDists, [νDists], noiseDist, signalModels)

Generate noisy data from unknown (and possibly known) parameter distributions.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use

N::Integer: Number of data points

xDists: Distributions of latent parameters [L] or scalar (if L = 1); xDists can be any object such that rand(xDists, ::Integer) is defined (or a collection of such objects)

νDists: Distributions of known parameters [K] or scalar (if K = 1); νDists can be any object such that rand(νDists, ::Integer) is defined (or a collection of such objects); omit this parameter if K = 0

noiseDist: Distribution of noise (assumes same noise distribution for both real and imaginary channels in complex case); noiseDist can be any object such that rand(noiseDist, ::Integer) is defined

signalModels::Union{<:Function,<:AbstractVector{<:Function}}: Signal models used to generate noiseless data [numSignalModels]; each signal model accepts as inputs L latent parameters (scalars) first, then K known parameters (scalars); user-defined parameters (e.g., scan parameters in MRI) should be built into the signal model

Return

y::Union{<:AbstractVector{<:Number},<:AbstractMatrix{<:Number}}: Output data of all the (simulated) signals [D,N] or [N] (if D = 1)

x::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Randomly generated latent parameters [L,N] or [N] (if L = 1)

ν::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Randomly generated known parameters [K,N] or [N] (if K = 1); not returned if νDists is omitted

PERK.holdout — Method

holdout([rng], N, T, λvals, ρvals, [weights,] xDistsTest, xDistsTrain,
        noiseDist, signalModels, kernelgenerator; showprogress)
holdout([rng], N, T, λvals, ρvals, [weights,] xDistsTest, νDistsTest,
        xDistsTrain, νDistsTrain, noiseDist, signalModels, kernelgenerator;
        showprogress)

Select λ and ρ via a holdout process.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use

N::Integer: Number of test points

T::Integer: Number of training points

λvals::AbstractVector{<:Real}: Values of λ to search over [nλ]

ρvals::AbstractVector{<:Real}: Values of ρ to search over [nρ]

weights::AbstractVector{<:Real}: Weights for calculating holdout cost [L]; omit if L = 1

xDistsTest: Distributions of latent parameters [L] or scalar (if L = 1); xDists can be any object such that rand(xDists, ::Integer) is defined (or a collection of such objects)

νDistsTest: Distributions of known parameters [K] or scalar (if K = 1); νDists can be any object such that rand(νDists, ::Integer) is defined (or a collection of such objects); omit this parameter if K = 0

xDistsTrain: Distributions of latent parameters [L] or scalar (if L = 1); xDists can be any object such that rand(xDists, ::Integer) is defined (or a collection of such objects)

νDistsTrain: Distributions of known parameters [K] or scalar (if K = 1); νDists can be any object such that rand(νDists, ::Integer) is defined (or a collection of such objects); omit this parameter if K = 0

noiseDist: Distribution of noise (assumes same noise distribution for both real and imaginary channels in complex case); noiseDist can be any object such that rand(noiseDist, ::Integer) is defined

signalModels::Union{<:Function,<:AbstractVector{<:Function}}: Signal models used to generate noiseless data [numSignalModels]; each signal model accepts as inputs L latent parameters (scalars) first, then K known parameters (scalars); user-defined parameters (e.g., scan parameters in MRI) should be built into the signal model

kernelgenerator::Function: Function that creates a Kernel object given a vector Λ of lengthscales

showprogress::Bool = false: Whether to show progress

Note

L is the number of unknown or latent parameters to be estimated

K is the number of known parameters

nλ is the number of λ values to try

nρ is the number of ρ values to try

Return

λ::Real: Bandwidth scaling parameter

ρ::Real: Tikhonov regularization parameter

Ψ::AbstractMatrix{<:Real}: Holdout costs for λvals and ρvals [nλ,nρ]

PERK.krr — Method

krr(ytest, trainData, kernel)

Predict latent parameters that generated ytest using kernel ridge regression.

Arguments

ytest::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Observed test data [Q,N] or [N] (if Q = 1) or scalar (if Q = N = 1)

trainData::TrainingData: Training data

kernel::Kernel: Kernel to use

Notes

Q is the number of observed features per test sample

N is the number of test samples

Return

xhat::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Estimated latent parameters [L,N] or [N] (if L = 1) or [L] (if N = 1) or scalar (if L = N = 1)

PERK.krr_train — Method

krr_train([rng], xtrain, ytrain, kernel, ρ)
krr_train(xtrain, ytrain, kernel, ρ, f, phase)

Train kernel ridge regression.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use (only used when kernel isa RFFKernel)

xtrain::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Latent parameters for training data [L,T] or [T] (if L = 1)

ytrain::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Features for training data [Q,T] or [T] (if Q = 1)

kernel::Kernel: Kernel to use

ρ::Real: Tikhonov regularization parameter

f::Union{<:AbstractVector{<:Real},AbstractMatrix{<:Real}} = randn(kernel.H, Q): Unscaled random frequency values [H,Q] or [H] (if Q = 1) (used when kernel isa RFFKernel)

phase::AbstractVector{<:Real} = rand(kernel.H): Random phase values [H] (used when kernel isa RFFKernel)

Note

L is the number of unknown or latent parameters to be predicted

Q is the number of observed features per training sample

T is the number of training samples

H is approximation order for kernels that use random Fourier features

Return

trainData::TrainingData: TrainingData object to be passed to krr

PERK.perk — Method

perk([rng], y, T, xDists, noiseDist, signalModels, kernel, ρ)
perk([rng], y, ν, T, xDists, νDists, noiseDist, signalModels, kernel, ρ)

Train PERK and then estimate latent parameters.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use

y::Union{<:Number,<:AbstractVector{<:Number},<:AbstractMatrix{<:Number}}: Test data points [D,N] or [N] (if D = 1) or scalar (if D = N = 1)

ν::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Known parameters [K,N] or [N] (if K = 1) or scalar (if K = N = 1); omit this parameter if K = 0

T::Integer: Number of training points

xDists: Distributions of latent parameters [L] or scalar (if L = 1); xDists can be any object such that rand(xDists, ::Integer) is defined (or a collection of such objects)

νDists: Distributions of known parameters [K] or scalar (if K = 1); νDists can be any object such that rand(νDists, ::Integer) is defined (or a collection of such objects); omit this parameter if K = 0

noiseDist: Distribution of noise (assumes same noise distribution for both real and imaginary channels in complex case); noiseDist can be any object such that rand(noiseDist, ::Integer) is defined

signalModels::Union{<:Function,<:AbstractVector{<:Function}}: Signal models used to generate noiseless data [numSignalModels]; each signal model accepts as inputs L latent parameters (scalars) first, then K known parameters (scalars); user-defined parameters (e.g., scan parameters in MRI) should be built into the signal model

kernel::Kernel: Kernel to use

ρ::Real: Tikhonov regularization parameter

Note

D is the combined number of outputs from all signal models (e.g., number of scans in MRI)

K is the number of known parameters

L is the number of unknown or latent parameters to be estimated

N is the number of test points (e.g., number of voxels in MRI)

Return

xhat::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Estimated latent parameters [L,N] or [N] (if L = 1) or [L] (if N = 1) or scalar (if L = N = 1)

PERK.perk — Method

perk(y, trainData, kernel)
perk(y, ν, trainData, kernel)

Estimate latent parameters using the provided training data.

Arguments

y::Union{<:Number,<:AbstractVector{<:Number},<:AbstractMatrix{<:Number}}: Test data points [D,N] or [N] (if D = 1) or scalar (if D = N = 1)

ν::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Known parameters [K,N] or [N] (if K = 1) or scalar (if K = N = 1); omit this parameter if K = 0

trainData::TrainingData: Training data

kernel::Kernel: Kernel to use

Return

xhat::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Estimated latent parameters [L,N] or [N] (if L = 1) or [L] (if N = 1) or scalar (if L = N = 1)

PERK.rffmap — Method

rffmap(q, freq, phase)

Map features to a higher dimensional space via random Fourier features.

Arguments

q::Union{<:Real,<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Lower-dimensional features [Q,N] or [N] (if Q = 1) or scalar (if Q = N = 1)

freq::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}: Random frequency values [H,Q] or [H] (if Q = 1)

phase::AbstractVector{<:Real}: Random phase values [H]

Note

Q is the number of features

N is the number of feature vectors in the input

H is the approximation order for the random Fourier features

Return

z::Union{<:AbstractVector{<:Real},<:AbstractMatrix{<:Real}}: Higher-dimensional features [H,N] or [H] (if N = 1)

PERK.train — Method

train([rng], T, xDists, [νDists], noiseDist, signalModels, kernel, ρ)

Train PERK using simulated training data.

Arguments

rng::AbstractRNG = Random.GLOBAL_RNG: Random number generator to use

T::Integer: Number of training points

xDists: Distributions of latent parameters [L] or scalar (if L = 1); xDists can be any object such that rand(xDists, ::Integer) is defined (or a collection of such objects)

νDists: Distributions of known parameters [K] or scalar (if K = 1); νDists can be any object such that rand(νDists, ::Integer) is defined (or a collection of such objects); omit this parameter if K = 0

noiseDist: Distribution of noise (assumes same noise distribution for both real and imaginary channels in complex case); noiseDist can be any object such that rand(noiseDist, ::Integer) is defined

signalModels::Union{<:Function,<:AbstractVector{<:Function}}: Signal models used to generate noiseless data [numSignalModels]; each signal model accepts as inputs L latent parameters (scalars) first, then K known parameters (scalars); user-defined parameters (e.g., scan parameters in MRI) should be built into the signal model

kernel::Kernel: Kernel to use

ρ::Real: Tikhonov regularization parameter

Return

trainData::TrainingData: TrainingData object to be passed to perk